pyproject.toml
[build-system]
requires    = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chuk-ai-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-session-manager>=0.1.3",
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "chuk-tool-processor>=0.1.2",
    "geopy>=2.4.1",
    "lxml-html-clean>=0.4.2",
    "openai>=1.76.2",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "readability-lxml>=0.8.1",
    "requests>=2.32.3",
    "uuid>=1.30",
]

# Tell setuptools to look in src/ for your a2a package
[tool.setuptools.packages.find]
where   = ["src"]
include = ["chuk_ai_planner*"]

# pytest settings so it finds your src/ layout automatically
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "strict"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
